+++ src/selection/TopNActivations.py
@@ -10,7 +10,7 @@ class TopNActivations(Selection):
         super(TopNActivations, self).__init__()
         self.hooks = []
         self.num_choices = num_choices
-        self.chosen_param_list = np.zeros(num_choices, dtype="int32")
+        self.chosen_param_list = np.zeros(num_choices, dtype=int)
         self.current = 0
 
         self.net = net
@@ -107,7 +107,7 @@ class TopNActivations(Selection):
             hook.remove()
 
     def initialize_neurons(self):
-        self.chosen_param_list = np.zeros(self.num_choices, dtype="int32")
+        self.chosen_param_list = np.zeros(self.num_choices, dtype=int)
         self.current = 0
 
     def _is_single_layer(self, module):
