+++ src/selection/TopNGradients.py
@@ -9,7 +9,7 @@ class TopNGradients(Selection):
     def __init__(self, net, num_choices):
         self.hooks = []
         self.num_choices = num_choices
-        self.chosen_param_list = np.zeros(num_choices, dtype="int32")
+        self.chosen_param_list = np.zeros(num_choices, dtype=int)
         self.current = 0
         self.require_backward = True
 
@@ -113,7 +113,7 @@ class TopNGradients(Selection):
             hook.remove()
 
     def initialize_neurons(self):
-        self.chosen_param_list = np.zeros(self.num_choices, dtype="int32")
+        self.chosen_param_list = np.zeros(self.num_choices, dtype=int)
         self.current = 0
 
     def _compute_num_param(self):
